refactor: LadderCreator dip
