+++ src/main/java/ladder/domain/LadderCreator.java
@@ -1,20 +1,5 @@
 package ladder.domain;
 
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-
-public class LadderCreator {
-    private static final LadderLineCreator ladderLineCreator = new LadderLineCreator();
-
-    public static Ladder create(LadderInfo ladderInfo, DirectionCreator directionCreator) {
-        return new Ladder(createLadderLines(ladderInfo, directionCreator));
-    }
-
-    private static List<LadderLine> createLadderLines(LadderInfo ladderInfo, DirectionCreator directionCreator) {
-        int height = ladderInfo.getHeight();
-        return IntStream.range(0, height)
-                .mapToObj(i -> ladderLineCreator.create(ladderInfo.getPlayerCount(), directionCreator))
-                .collect(Collectors.toList());
-    }
+public interface LadderCreator {
+    Ladder create(LadderInfo ladderInfo, DirectionCreator directionCreator);
 }
