+++ src/test/java/ladder/domain/LadderCreatorTest.java
@@ -1,5 +1,6 @@
 package ladder.domain;
 
+import ladder.infra.BasicLadderCreator;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
@@ -11,7 +12,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 class LadderCreatorTest {
     @Test
     void create() {
-        Ladder ladder = LadderCreator.create(new LadderInfo(3, 2), () -> true);
+        Ladder ladder = BasicLadderCreator.create(new LadderInfo(3, 2), () -> true);
         assertThat(ladder).isEqualTo(new Ladder(Arrays.asList(
                 new LadderLine(Arrays.asList(
                         new Point(0, Direction.of(FALSE, TRUE)),
